make the preview-button name more generic
